iot-simulator/requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
web3==6.15.1
eth-account==0.11.0
eth-keys==0.4.0
eth-utils==2.3.1
